Location refactor
